pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chesspoint72"
version = "0.1.0"
description = "Pygame chess renderer with UCI engine support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pygame>=2.5.0",
  "python-chess>=1.999",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
chesspoint72 = "chesspoint72.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

